lab2 completed, all systems functional

+++ include/particlesystem.h
@@ -23,7 +23,7 @@ public:
 
     void addEffect(Effect* e);
 
-    int numberOfParticles();
+    size_t numberOfParticles();
 
     std::vector<Particle> _particles;
 private:
